Creating the helper for the new Solidity structure after adding payments
